// src/app/shared/models/routes-rp.ts
export enum RoutesRp {
  ActivityIntro = '/activity/intro',
  ActivityIntroPreview = '/activity/intro/preview',
  ActivityManagerTemplate = '/activity-manager/template',
  ActivityQuestionPreview = '/activity/question/preview', // Navigation to preview a singular question from questions list
  ActivityQuestions = '/activity/questions',
  ActivityQuestionsPreview = '/activity/questions/preview',
  EvaluationPreview = '/activity/evaluation/preview',
  Home = '/home'
}